@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <AssemblyVersion>1.0.0.3</AssemblyVersion>
+    <TargetFramework>net5.0</TargetFramework>
+    <AssemblyVersion>2.0.0.0</AssemblyVersion>
     <FileVersion>1.0.0.0</FileVersion>
     <NeutralLanguage>de-DE</NeutralLanguage>
     <Authors>Patrick Becker</Authors>
@@ -11,7 +11,13 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="HtmlAgilityPack" Version="1.8.9" />
+    <PackageReference Include="HtmlAgilityPack" Version="1.11.34" />
   </ItemGroup>
+
+   <ItemGroup>
+    <None Update="banner.template">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+   </ItemGroup>
 
 </Project>
